Fixed issues in events

+++ functions.h
@@ -9,7 +9,7 @@ void generate_board(Cell board[]);
 void initialize_players(Player players[]);
 void generate_event_cards(Events national_events[]);
 void destroy_property(Cell *place);
-int round_off(float num);
+int round_off(double num);
 void sort_players(Player players[]);
 void decide_player_order(Player players[]);
